plattenbestand/requirements.txt
Flask==3.1.1
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.2
Flask-Migrate==4.1.0
Werkzeug==3.1.3
SQLAlchemy==2.0.36
openpyxl==3.1.5
psycopg2-binary==2.9.10
bcrypt==4.2.1
gunicorn==23.0.0
python-dotenv==1.0.1
